useResize hook, updated tests, minor fixes
