+++ package.json
@@ -23,6 +23,7 @@
 		"@testing-library/dom": "^10.4.0",
 		"@testing-library/jest-dom": "^6.5.0",
 		"@testing-library/react": "^16.0.1",
+		"@testing-library/user-event": "^14.5.2",
 		"@types/jest": "^29.5.13",
 		"@types/node": "^22.5.5",
 		"@types/react": "^18.3.3",
